diffrent daily pickup methods
