+++ src/main.ts
@@ -8,7 +8,7 @@ import { GetNmSecrets } from './utility/nm-secrets.utility';
 import { ConfigInstanceByGuildIdGet, InitInstanceServices } from './utility';
 import { type GuildServiceMapModel } from './model';
 import { AddCron } from './utility/cron-utility';
-import { FoodCountReminder, DailyPickupsThread } from './jobs';
+import { FoodCountReminder, DailyPickupsWithoutThread } from './jobs';
 
 const GuildServiceMap: GuildServiceMapModel = {};
 
@@ -45,7 +45,7 @@ async function main() {
         }
 
         for (const guild of c.guilds.cache.values()) {
-            DailyPickupsThread(guild, GuildServiceMap[guild.id])
+            DailyPickupsWithoutThread(guild, GuildServiceMap[guild.id])
         }
     });
